requirements.txt
scipy==1.5.0
pydicom==2.0.0
numpy==1.19.0
dcmstack==0.8.0
coloredlogs==14.0
opencv_python==4.2.0.34
matplotlib==3.2.2
fire==0.3.1
scikit_image==0.17.2
Keras==2.2.4
nibabel==3.1.0
Pillow==7.2.0
SimpleITK==2.0.1
tensorflow==1.13.1
cffi==1.14.1
unwrap==0.1.1
h5py==2.10.0
